pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "production-agents"
version = "0.1.0"
description = "Build Production-Ready AI Agents - O'Reilly Live Learning"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Tim Warner", email = "timothywarner@example.com"}
]
keywords = [
    "ai-agents",
    "langgraph",
    "crewai",
    "mcp",
    "claude",
    "llm",
    "autonomous-agents"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.2.0",
    "langchain-openai>=0.2.0",
    "crewai>=0.80.0",
    "crewai-tools>=0.14.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "rich>=13.9.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "black>=24.10.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]
db = [
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.29.0",
]
api = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
]

[project.urls]
Homepage = "https://github.com/timothywarner-org/production-agents"
Documentation = "https://github.com/timothywarner-org/production-agents#readme"
Repository = "https://github.com/timothywarner-org/production-agents.git"
Issues = "https://github.com/timothywarner-org/production-agents/issues"

[project.scripts]
verify-setup = "scripts.verify_setup:main"
cost-calc = "scripts.cost_calculator:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["segment_*", "projects*", "scripts*"]

[tool.black]
line-length = 88
target-version = ["py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"
